Sync from downstream

+++ index.js
@@ -3,7 +3,6 @@ const url = require("url");
 const https = require("https");
 const log = (...args) => console.log("→", ...args);
 const list = require("./videojson.js");
-const promises = [];
 
 function loadVideo(num, cb) {
   let masterUrl = list[num].url;
@@ -11,7 +10,7 @@ function loadVideo(num, cb) {
     masterUrl += "?base64_init=1";
   }
 
-  getJson(masterUrl, (err, json) => {
+  getJson(masterUrl, num, (err, json) => {
     if (err) {
       return cb(err);
     }
@@ -40,7 +39,7 @@ function loadVideo(num, cb) {
       list[num].name + ".m4v",
       err => {
         if (err) {
-          return cb(err);
+          cb(err);
         }
 
         processFile(
@@ -51,7 +50,7 @@ function loadVideo(num, cb) {
           list[num].name + ".m4a",
           err => {
             if (err) {
-              return cb(err);
+              cb(err);
             }
 
             cb(null, num + 1);
@@ -63,14 +62,15 @@ function loadVideo(num, cb) {
 }
 
 function processFile(type, baseUrl, initData, segments, filename, cb) {
-  const filePath = `./parts/${filename}`;
-  const downloadingFlag = `./parts/.${filename}~`;
-  
-  if(fs.existsSync(downloadingFlag)) {
-    log("⚠️", ` ${filename} - ${type} is incomplete, restarting the download`);
+  const file = filename.replace(/[^\w.]/gi, '-');
+  const filePath = `./parts/${file}`;
+  const downloadingFlag = `./parts/.${file}~`;
+
+  if (fs.existsSync(downloadingFlag)) {
+    log("⚠️", ` ${file} - ${type} is incomplete, restarting the download`);
   } else if (fs.existsSync(filePath)) {
-    log("⚠️", ` ${filename} - ${type} already exists`);
-    return cb();
+    log("⚠️", ` ${file} - ${type} already exists`);
+    cb();
   } else {
     fs.writeFileSync(downloadingFlag, '');
   }
@@ -96,7 +96,9 @@ function processFile(type, baseUrl, initData, segments, filename, cb) {
 
 function combineSegments(type, i, segmentsUrl, output, filename, downloadingFlag, cb) {
   if (i >= segmentsUrl.length) {
-    fs.unlinkSync(downloadingFlag);
+    if (fs.existsSync(downloadingFlag)) {
+      fs.unlinkSync(downloadingFlag);
+    }
     log("🏁", ` ${filename} - ${type} done`);
     return cb();
   }
@@ -125,17 +127,20 @@ function combineSegments(type, i, segmentsUrl, output, filename, downloadingFlag
   });
 }
 
-function getJson(url, cb) {
+function getJson(url, n, cb) {
   let data = "";
 
   https
     .get(url, res => {
-      res.on("data", d => (data += d));
-
-      res.on("end", () => cb(null, JSON.parse(data)));
+      if (res.statusMessage.toLowerCase() !== 'gone') {
+        res.on("data", d => (data += d));
+        res.on("end", () => cb(null, JSON.parse(data)));
+      } else {
+        return cb(`The master.json file is expired or crushed. Please update or remove it from the sequence (broken on ` + n + ` position)`);
+      }
     })
     .on("error", e => {
-      cb(e);
+      return cb(e);
     });
 }
 
@@ -145,7 +150,6 @@ function initJs(n = 0) {
   loadVideo(n, (err, num) => {
     if (err) {
       log("⚠️", ` ${err}`);
-      return;
     }
 
     if (list[num]) {
